pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opinion-trade-sdk"
version = "0.1.0"
description = "Opinion Trade Python SDK"
authors = [{ name = "Opinion Trade" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
    "eth-account>=0.10.0,<0.13",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["opinion_trade_sdk", "opinion_trade_sdk.*"]
